A few updates. refactor code and update ug,dg. add ppp
