--- docs/team/johndoe.md
@@ -1,6 +1,0 @@
-# John Doe - Project Portfolio Page
-
-## Overview
-
-
-### Summary of Contributions
